backend/requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
orjson==3.10.7
python-dotenv==1.0.1
SQLAlchemy==2.0.32
aiosqlite==0.20.0
asyncpg==0.29.0
psycopg[binary,pool]==3.2.1
databases[postgresql]==0.9.0
redis==5.0.8
websockets==12.0
protobuf==5.27.2
ortools==9.10.4067
torch==2.3.1
ray[rllib]==2.34.0
networkx==3.3
numpy==1.26.4
pandas==2.2.2
tenacity==8.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
matplotlib==3.8.4
PyPDF2==3.0.1
Pillow==10.4.0
pytesseract==0.3.10

